refactor: crashReporterInit
